build script now generates GetInfo and GetBlockChainInfo json files which are parsed by proc-macro to generate Response struct types
