+++ Cargo.toml
@@ -14,7 +14,9 @@ zcashrpc-macros = { path = "zcashrpc-macros" }
 tokio = { version = "*", features = ["macros"] }
 reqwest = { version = "0.10.8", features = ["json"] }
 serde = { version = "*", features = ["derive"] }
+serde_derive = "*"
 serde_json = "*"
+json_typegen = "0.4.1"
 derive_more = "*"
 base64 = "*"
 eyre = "*"
@@ -29,7 +31,6 @@ reqwest = { version = "0.10.8", features = ["json"] }
 serde = { version = "*", features = ["derive"] }
 dirs = { version = "*", optional = true }
 serde_json = "*"
-json_typegen = "0.4.1"
 base64 = "*"
