+++ zcashrpc-macros/Cargo.toml
@@ -10,3 +10,4 @@ proc-macro = true
 [dependencies]
 quote = "1.0"
 proc-macro2 = "1.0"
+json_typegen = "0.4.1"
